Rename unit tests for automatic discovery
