+++ ndex/test/test_network.py
