+++ ndex/test/test_properties.py
