+++ ndex/test/test_query.py
