+++ ndex/test/test_setting_visibility.py
