+++ ndex/test/test_system_properties.py
